backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.2
openai>=1.3.0
python-dotenv>=1.0.0
jinja2>=3.1.2
pydantic>=2.4.2 